refactor: 지도 좌표 부동 소수점 수정및 주문 횟수와 주문 안의 상품 주문 갯수 분리

+++ src/main/java/com/jigume/dto/member/MemberInfoDto.java
@@ -12,9 +12,9 @@ public class MemberInfoDto {
 
     private String profileImgUrl;
 
-    private Long mapX;
+    private Double mapX;
 
-    private Long mapY;
+    private Double mapY;
 
     //TODO: 도로명 주소
